requirements.txt
msal==1.26.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
pytest==7.4.3